chore: remove unused file utils
